cekura-mcp-server/requirements.txt
# Core dependencies
fastmcp>=0.1.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
